beauty_backend/requirements.txt
# API Framework
fastapi==0.115.12
uvicorn[standard]==0.32.1
python-multipart>=0.0.20    # Required for file uploads (UploadFile)
pydantic==2.10.3
pydantic-settings==2.7.0

# Security
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.9

# ML Inference
onnxruntime==1.16.0
scikit-learn==1.3.2

# Image Processing
Pillow==11.0.0
scikit-image==0.24.0  # For validated RGB→LAB conversion

# Database
sqlalchemy>=2.0.0
mysql-connector-python>=9.0.0   # For mysql+mysqlconnector:// URLs
pymysql>=1.0.0                  # Alternative MySQL driver (mysql+pymysql:// URLs)

# Utilities
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
